Chat-bot StockSense/requirements.txt
faiss-cpu>=1.8.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
pypdf>=4.2.0
pymongo>=4.7.0
python-dotenv>=1.0.1
requests>=2.31.0
sentence-transformers>=2.7.0
